+++ plugins/kubernetes-backend/src/types/types.ts
@@ -104,6 +104,8 @@ export type KubernetesObjectTypes =
   | 'customresources'
   | 'statefulsets'
   | 'daemonsets';
+// If updating this list, also make sure to update
+// `objectTypes` and `apiVersionOverrides` in config.d.ts!
 
 /**
  * Used to load cluster details from different sources