+++ plugins/scaffolder-react/src/next/components/Stepper/Stepper.test.tsx
@@ -237,6 +237,50 @@ describe('Stepper', () => {
     expect(getByText('im a custom field extension')).toBeInTheDocument();
   });
 
+  it('should disable the form with progress when async validators are running', async () => {
+    const MockComponent = () => {
+      return <h1>im a custom field extension</h1>;
+    };
+
+    const manifest: TemplateParameterSchema = {
+      title: 'Custom Fields',
+      steps: [
+        {
+          title: 'Test',
+          schema: {
+            properties: {
+              name: {
+                type: 'string',
+                'ui:field': 'Mock',
+              },
+            },
+          },
+        },
+      ],
+    };
+
+    const { getByRole } = await renderInTestApp(
+      <Stepper
+        manifest={manifest}
+        extensions={[
+          {
+            name: 'Mock',
+            component: MockComponent,
+            validation: async () => new Promise(r => setTimeout(r, 1000)),
+          },
+        ]}
+        onCreate={jest.fn()}
+      />,
+    );
+
+    await act(async () => {
+      await fireEvent.click(getByRole('button', { name: 'Review' }));
+
+      expect(getByRole('progressbar')).toBeInTheDocument();
+      expect(getByRole('button', { name: 'Review' })).toBeDisabled();
+    });
+  });
+
   it('should transform default error message', async () => {
     const manifest: TemplateParameterSchema = {
       steps: [