+++ packages/core-components/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @backstage/core-components
 
+## 0.13.0-next.3
+
+### Minor Changes
+
+- 01cd4e25754: **BREAKING:** Removing `Tabs` component from `core-components` as it is neither used in the core Backstage app nor in the monorepo plugins. If you are using this component in your instance please consider replacing it with the [Material UI `Tabs`](https://v4.mui.com/components/tabs/#tabs) component like the following:
+
+  ```diff
+  - <Tabs
+  -    tabs={[{
+  -       icon: <AccessAlarmIcon />,
+  -       content: <div>Label</div>,
+  -    }]}
+  - />
+
+  + <Tabs>
+  +    <Tab
+  +       label = "Label"
+  +       icon = { <AccessAlarmIcon /> }
+  +    />
+  + </Tabs>
+  ```
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/config@1.0.7
+  - @backstage/core-plugin-api@1.5.1-next.1
+  - @backstage/errors@1.1.5
+  - @backstage/theme@0.2.19-next.0
+  - @backstage/version-bridge@1.0.4-next.0
+
 ## 0.12.6-next.2
 
 ### Patch Changes