+++ plugins/catalog-graphql/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-catalog-graphql
 
+## 0.3.20-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/catalog-model@1.3.0-next.0
+  - @backstage/config@1.0.7
+  - @backstage/types@1.0.2
+
 ## 0.3.20-next.1
 
 ### Patch Changes