+++ plugins/config-schema/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage/plugin-config-schema
 
+## 0.1.40-next.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.13.0-next.3
+  - @backstage/config@1.0.7
+  - @backstage/core-plugin-api@1.5.1-next.1
+  - @backstage/errors@1.1.5
+  - @backstage/theme@0.2.19-next.0
+  - @backstage/types@1.0.2
+
 ## 0.1.40-next.2
 
 ### Patch Changes