+++ plugins/scaffolder-react/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @backstage/plugin-scaffolder-react
 
+## 1.3.0-next.3
+
+### Patch Changes
+
+- d2488f5e54c: Add indication that the validators are running
+- 8c40997df44: Updated dependency `@rjsf/core-v5` to `npm:@rjsf/core@5.5.2`.
+- Updated dependencies
+  - @backstage/plugin-catalog-react@1.5.0-next.3
+  - @backstage/catalog-model@1.3.0-next.0
+  - @backstage/core-components@0.13.0-next.3
+  - @backstage/catalog-client@1.4.1-next.1
+  - @backstage/core-plugin-api@1.5.1-next.1
+  - @backstage/errors@1.1.5
+  - @backstage/theme@0.2.19-next.0
+  - @backstage/types@1.0.2
+  - @backstage/version-bridge@1.0.4-next.0
+  - @backstage/plugin-scaffolder-common@1.2.7-next.2
+
 ## 1.3.0-next.2
 
 ### Patch Changes