+++ plugins/search-backend-module-elasticsearch/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage/plugin-search-backend-module-elasticsearch
 
+## 1.2.0-next.3
+
+### Patch Changes
+
+- 104b6b19487: fix ElasticSearch throwing error when index is missing
+- Updated dependencies
+  - @backstage/backend-common@0.18.4-next.2
+  - @backstage/backend-plugin-api@0.5.1-next.2
+  - @backstage/config@1.0.7
+  - @backstage/plugin-search-backend-node@1.2.0-next.2
+  - @backstage/plugin-search-common@1.2.3-next.0
+
 ## 1.2.0-next.2
 
 ### Patch Changes